kernel/scheduler: Fix core selection, quantum scaling, and race conditions
This commit addresses several issues identified in the Haiku kernel scheduler:

1. low_latency.cpp: Home package selection now respects thread coloring
   (P-core vs E-core preference) for high/low priority threads.
2. scheduler_thread.cpp: Fixed potential 32-bit overflow in the interactivity
   multiplier calculation by using explicit 64-bit casts.
3. scheduler_topology.h: Increased search attempts in search_global_random
   to mitigate collisions on systems with many packages.
4. scheduler_thread.h: Added a NULL check for fCore in _UpdatePriorityBoost
   to prevent crashes during transient migration states.
5. RunQueue.h: Implemented an adaptive search limit in PeekOption to ensure
   the work-stealing budget is not exhausted at high-priority levels before
   reaching lower-priority levels.
6. scheduler.cpp: Linked interaction update throttling to MinimalQuantum
   and added a retry loop for atomic updates to prevent silent drops.
7. power_saving.cpp: Added defensive NULL checks for Package() and Node()
   in the IRQ rebalancing path to prevent crashes during hot-unplug.

+++ src/system/kernel/scheduler/scheduler_thread.cpp
@@ -367,7 +367,7 @@ ThreadData::ComputeQuantum() const
 	// Context-aware quantum scaling: scale by interactivity score (0.5x - 1.5x)
 	// Fast integer approximation of / 1000 (1049 / 2^20 ~= 0.0010004)
 	// Ensure 64-bit arithmetic to prevent overflow.
-	quantum = ((int64)quantum * (1500 - fInteractivityScore) * 1049) >> 20;
+	quantum = (int64)quantum * (int64)(1500 - fInteractivityScore) * 1049 >> 20;
 
 	// Clamp to [floor, maxAllowed].
 	// Lower bound: the interactivity multiplier (0.5x at fInteractivityScore=1000)
